completed config DB

+++ src/main/java/com/example/imageconventer/service/LoginService.java
@@ -3,5 +3,7 @@ package com.example.imageconventer.service;
 import com.example.imageconventer.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginService extends JpaRepository<User, Integer> {
+public interface LoginService extends JpaRepository<User, Long> {
+
 }
+
